SchoolManagementBackend/CollegeManagement/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0

# Production dependencies
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
dj-database-url==2.1.0
requests==2.32.3
